models/denoiser/requirements_cuda.txt
hydra_core==0.11.3
hydra_colorlog==0.1.4
numpy
pystoi==0.3.3
git+https://github.com/ludlows/python-pesq#egg=pesq
six
sounddevice==0.4.0
torchaudio
torch
